pyproject.toml
[project]
name = "django-cotton-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cotton>=1.2.1",
    "nanodjango>=0.9.2",
]
